app/requirements.txt
google
google-genai
dotenv
pandas
numpy
Flask
faiss-cpu
